# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KBJ API",
    "description": "KBJ 로그인 등급 API(사용자 1명). 모든 데이터 응답은 봉투(source·as_of·quality·notes·generated_at·data), 금액은 원 단위 정수. 외부 원천을 부르지 않는다(DB·Redis 읽기만).",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/market/summary": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Summary",
        "operationId": "summary_api_market_summary_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_MarketSummary_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/sectors": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Sectors",
        "operationId": "sectors_api_market_sectors_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                1,
                5,
                20
              ],
              "type": "integer",
              "default": 1,
              "title": "Period"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_SectorHeat_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/market/ribbon": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Ribbon",
        "operationId": "ribbon_api_market_ribbon_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Ribbon_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/flows/screen": {
      "get": {
        "tags": [
          "flows"
        ],
        "summary": "Screen",
        "operationId": "screen_api_flows_screen_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "value",
                "foreign",
                "inst",
                "both",
                "streak",
                "spike"
              ],
              "type": "string",
              "title": "Mode"
            }
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "KOSPI",
                "KOSDAQ"
              ],
              "type": "string",
              "default": "all",
              "title": "Market"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                1,
                5,
                20
              ],
              "type": "integer",
              "default": 5,
              "title": "Period"
            }
          },
          {
            "name": "min_avg_turnover",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1000000000000000,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Avg Turnover"
            }
          },
          {
            "name": "include_flagged",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Flagged"
            }
          },
          {
            "name": "share_class",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "common",
                "pref"
              ],
              "type": "string",
              "default": "common",
              "title": "Share Class"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ScreenResponse_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/investors": {
      "get": {
        "tags": [
          "flows"
        ],
        "summary": "Investors",
        "operationId": "investors_api_flows_investors_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_InvestorTotals_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/stock/{code}": {
      "get": {
        "tags": [
          "flows"
        ],
        "summary": "Stock",
        "operationId": "stock_api_flows_stock__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9A-Z]{6}$",
              "title": "Code"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 60,
              "minimum": 1,
              "default": 20,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_StockFlowDetail_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/flows/intraday": {
      "get": {
        "tags": [
          "flows"
        ],
        "summary": "Intraday",
        "operationId": "intraday_api_flows_intraday_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_IntradayFlows_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_IntradayFlows_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/IntradayFlows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[IntradayFlows]"
      },
      "Envelope_InvestorTotals_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/InvestorTotals"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[InvestorTotals]"
      },
      "Envelope_MarketSummary_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/MarketSummary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[MarketSummary]"
      },
      "Envelope_Ribbon_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/Ribbon"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[Ribbon]"
      },
      "Envelope_ScreenResponse_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/ScreenResponse"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[ScreenResponse]"
      },
      "Envelope_SectorHeat_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/SectorHeat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[SectorHeat]"
      },
      "Envelope_StockFlowDetail_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/StockFlowDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[StockFlowDetail]"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorBody",
        "description": "오류 응답(401·403·413·415·503) — 사유 코드만, 입력값·내부 문구 없음."
      },
...
      "MarketSummary": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "indices": {
            "items": {
              "$ref": "#/components/schemas/IndexTile"
            },
            "type": "array",
            "title": "Indices"
          },
          "turnover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TurnoverPanel"
              },
              {
                "type": "null"
              }
            ]
          },
          "breadth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Breadth"
              },
              {
                "type": "null"
              }
            ]
          },
          "investors": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InvestorTotals"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "indices",
          "turnover",
          "breadth",
          "investors"
        ],
        "title": "MarketSummary"
      },
...
      "NoDataBody": {
        "properties": {
          "code": {
            "type": "string",
            "const": "no_data",
            "title": "Code",
            "default": "no_data"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "NoDataBody",
        "description": "데이터가 아직 없을 때(404) — 화면은 '아직 없음 — <작업> <예정 시각>' 으로 그린다(§6.7)."
      },
...
      "Quality": {
        "type": "string",
        "enum": [
          "ok",
          "stale",
          "estimated",
          "invalid"
        ],
        "title": "Quality",
        "description": "값의 품질. 문자열 값이 그대로 DB `quality` 열과 API 응답에 들어간다."
      },
...
      "Ribbon": {
        "properties": {
          "chips": {
            "items": {
              "$ref": "#/components/schemas/Chip"
            },
            "type": "array",
            "title": "Chips"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "chips"
        ],
        "title": "Ribbon"
      },
...
      "ScreenResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "value",
              "foreign",
              "inst",
              "both",
              "streak",
              "spike"
            ],
            "title": "Mode"
          },
          "market": {
            "type": "string",
            "enum": [
              "all",
              "KOSPI",
              "KOSDAQ"
            ],
            "title": "Market"
          },
          "period": {
            "type": "integer",
            "title": "Period"
          },
          "share_class": {
            "type": "string",
            "enum": [
              "common",
              "pref"
            ],
            "title": "Share Class"
          },
          "min_avg_turnover": {
            "type": "integer",
            "title": "Min Avg Turnover"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "n_total": {
            "type": "integer",
            "title": "N Total"
          },
          "n_excluded": {
            "$ref": "#/components/schemas/Excluded"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ScreenRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "market",
          "period",
          "share_class",
          "min_avg_turnover",
          "date",
          "n_total",
          "n_excluded",
          "rows"
        ],
        "title": "ScreenResponse"
      },
...
      "SectorHeat": {
        "properties": {
          "period": {
            "type": "integer",
            "title": "Period"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/SectorCell"
            },
            "type": "array",
            "title": "Cells"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period",
          "cells"
        ],
        "title": "SectorHeat"
      },
...
      "StockFlowDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/DayFlow"
            },
            "type": "array",
            "title": "Days"
          },
          "cumulative": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Cumulative"
          },
          "checks": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Checks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name",
          "days",
          "cumulative",
          "checks"
        ],
        "title": "StockFlowDetail"
      },
...
    }
  }
}